refactor: reply 도메인 패키지 구조 및 클래스명 변경
