+++ src/main/java/online/bottler/reply/adaptor/in/web/ReplyController.java
@@ -1,15 +1,15 @@
-package online.bottler.reply.presentation;
+package online.bottler.reply.adaptor.in.web;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import online.bottler.reply.application.port.in.ReplyUseCase;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import online.bottler.global.response.ApiResponse;
-import online.bottler.reply.application.dto.response.ReplyResponseDTO;
-import online.bottler.reply.application.ReplyService;
+import online.bottler.reply.application.response.ReplyResponse;
 import online.bottler.user.auth.CustomUserDetails;
 
 @RestController
@@ -17,11 +17,11 @@ import online.bottler.user.auth.CustomUserDetails;
 @RequestMapping("/reply")
 @Tag(name = "최근 답장 3개 보여주는 컨트롤러")
 public class ReplyController {
-    private final ReplyService replyService;
+    private final ReplyUseCase replyUseCase;
 
     @GetMapping
-    public ApiResponse<List<ReplyResponseDTO>> findRecentReplyLetters(
+    public ApiResponse<List<ReplyResponse>> findRecentReplyLetters(
             @AuthenticationPrincipal CustomUserDetails customUserDetails) {
-        return ApiResponse.onSuccess(replyService.findRecentReplyLetters(customUserDetails.getUserId()));
+        return ApiResponse.onSuccess(replyUseCase.findRecentReplyLetters(customUserDetails.getUserId()));
     }
 }
